refactoring for upcoming changes
