+++ .github/workflows/android-ad-interstitial.yml
@@ -10,7 +10,7 @@ on:
       - "CHANGELOG.md"
       - "scripts/version.py"
       - "tests/android_ci/buildozer.spec.in"
-      - "kivmob.py"
+      - "kivmob/**"
       - "bridges/android/**"
       - "scripts/ci/**"
       - "docker-compose.android.yml"
@@ -23,7 +23,7 @@ on:
       - "CHANGELOG.md"
       - "scripts/version.py"
       - "tests/android_ci/buildozer.spec.in"
-      - "kivmob.py"
+      - "kivmob/**"
       - "bridges/android/**"
       - "scripts/ci/**"
       - "docker-compose.android.yml"
