+++ kivmob/__main__.py
@@ -1,0 +1,11 @@
+"""``python -m kivmob`` entry point."""
+
+print(
+    "\033[92m  _  ___       __  __       _\n"
+    " | |/ (_)_   _|  \\/  | ___ | |__\n"
+    " | ' /| \\ \\ / / |\\/| |/ _ \\| '_ \\\n"
+    " | . \\| |\\ V /| |  | | (_) | |_) |\n"
+    " |_|\\_\\_| \\_/ |_|  |_|\\___/|_.__/\n\033[0m"
+)
+print(" AdMob support for Kivy\n")
+print(" Michael Stott, 2019\n")
