+++ kivmob/desktop.py
@@ -1,0 +1,9 @@
+"""Desktop (no-op) AdMob bridge."""
+
+from .bridge import AdMobBridge
+
+
+class DesktopBridge(AdMobBridge):
+    """No-op AdMob bridge for desktop (and Android when AdMob fails to load)."""
+
+    pass
