+++ kivmob/ios.py
@@ -1,0 +1,9 @@
+"""iOS AdMob bridge stub for future development."""
+
+from .bridge import AdMobBridge
+
+
+class iOSBridge(AdMobBridge):
+    """Stub AdMob bridge for future iOS support."""
+
+    pass
